prac-3/pyprac-3/requirements.txt
pysimplegui==4.60.4
Pillow==9.5.0
scipy==1.10.1
numpy==1.24.3